Clean up lucky egg logic to always check expiration.
