+++ PoGo.NecroBot.Logic/Inventory.cs
@@ -24,6 +24,8 @@ using TinyIoC;
 using static PoGo.NecroBot.Logic.Model.Settings.FilterUtil;
 using POGOProtos.Settings.Master.Pokemon;
 using PoGo.NecroBot.Logic.Logging;
+using PoGo.NecroBot.Logic.Event;
+using PoGo.NecroBot.Logic.Common;
 
 #endregion
 
@@ -36,6 +38,7 @@ namespace PoGo.NecroBot.Logic
         private GetPlayerResponse _player = null;
         private int _level = 0;
         private IEnumerable<PokemonSettings> _pokemonSettings = null;
+        private DateTime _lastLuckyEggTime;
 
         private readonly List<ItemId> _revives = new List<ItemId> { ItemId.ItemRevive, ItemId.ItemMaxRevive };
         private ISession ownerSession;
@@ -545,10 +548,56 @@ namespace PoGo.NecroBot.Logic
                 .Where(p => p != null);
         }
 
-        public async Task<UseItemXpBoostResponse> UseLuckyEggConstantly()
+        public async Task UseLuckyEgg()
         {
-            var UseLuckyEgg = await _client.Inventory.UseItemXpBoost().ConfigureAwait(false);
-            return UseLuckyEgg;
+            var inventoryContent = await ownerSession.Inventory.GetItems().ConfigureAwait(false);
+
+            var luckyEgg = inventoryContent.FirstOrDefault(p => p.ItemId == ItemId.ItemLuckyEgg);
+
+            if (luckyEgg == null || luckyEgg.Count == 0) // We tried to use egg but we don't have any more. Just return.
+            {
+                Logger.Write(ownerSession.Translation.GetTranslation(TranslationString.NoEggsAvailable));
+                return;
+            }
+
+            if (_lastLuckyEggTime.AddMinutes(30).Ticks > DateTime.Now.Ticks)
+            {
+                TimeSpan duration = _lastLuckyEggTime.AddMinutes(30) - DateTime.Now;
+                Logger.Write(ownerSession.Translation.GetTranslation(TranslationString.UseLuckyEggActive, duration.Minutes, duration.Seconds));
+                return;
+            }
+
+            var responseLuckyEgg = await ownerSession.Client.Inventory.UseItemXpBoost().ConfigureAwait(false);
+            switch (responseLuckyEgg.Result)
+            {
+                case UseItemXpBoostResponse.Types.Result.Success:
+                    {
+                        _lastLuckyEggTime = DateTime.Now;
+                        ownerSession.EventDispatcher.Send(new UseLuckyEggEvent { Count = luckyEgg.Count - 1 });
+                        TinyIoCContainer.Current.Resolve<MultiAccountManager>().DisableSwitchAccountUntil(DateTime.Now.AddMinutes(30));
+                        Logger.Write(ownerSession.Translation.GetTranslation(TranslationString.UsedLuckyEgg));
+                    }
+                    break;
+
+                case UseItemXpBoostResponse.Types.Result.ErrorNoItemsRemaining:
+                    {
+                        Logger.Write(ownerSession.Translation.GetTranslation(TranslationString.NoEggsAvailable));
+                    }
+                    break;
+
+                case UseItemXpBoostResponse.Types.Result.ErrorXpBoostAlreadyActive:
+                    {
+                        TimeSpan duration = _lastLuckyEggTime.AddMinutes(30) - DateTime.Now;
+                        Logger.Write(ownerSession.Translation.GetTranslation(TranslationString.UseLuckyEggActive, duration.Minutes, duration.Seconds));
+                    }
+                    break;
+
+                default:
+                    Logger.Write($"Failed to use a Lucky Egg!", LogLevel.Error);
+                    break;
+            }
+
+            await DelayingUtils.DelayAsync(ownerSession.LogicSettings.DelayBetweenPlayerActions, 0, ownerSession.CancellationTokenSource.Token).ConfigureAwait(false);
         }
 
         public async Task<UseIncenseResponse> UseIncenseConstantly()
